Inline ckbtc_minter_update_balance
