+++ rs/tests/src/ckbtc/minter/mod.rs
@@ -1,2 +1,1 @@
-pub mod test_update_balance;
 pub mod utils;
